Added insulation to the default modsuit chestpeice
The default (Non-Solarian) modsuit chestplate now provides temperature insulation without needing to seal the whole suit, to match a previous change made to the "Mazu" (Solarian) Modsuit chestplate. This should make the the modsuit fall more in line with its theme of being an adaptable utility suit, and encourage it's use in civilian departments where temperature or pressure could be an issue, but a dedicated hardsuit is impractical or undesirable.