Cobra sound fix
The cobra now actually makes a sound when fired, while still not being able to be heard unless youre close by. (Distance can be changed if it proves too small a range.)
On top of this I've swapped the sound the pistol uses to the sound the wspr uses, as it's higher quality, and more pleasant on the ears, while still being distinctly recognizable as a syndicate weapon by the sound.