Restored xeno ghost roles
Xenos are ghost roles again, along with minor damage buffs, so that they are less of an annoyance, and more of actual threat, without making them too powerful. An unarmored baseline human will take 7 hits to crit by most xenos. with the exceptions being 9 hits from a runner, roughly 4-5 from the queen, and 6 from a ravager.